Team 4/requirements.txt
#Python 3.11
black
boto3
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
pillow==11.0.0
pandas
pytest
pylint
pyyaml
pydot-env
scipy
seaborn
sympy==1.13.1
timm==0.4.12
tqdm
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2

#  need to add in boto3 awscli maybe
# tensorflow pydot-env pylint

# include in req
# !pip install pandas as pd
# !pip install python-dotenv
# !pip install pyyaml